requirements.txt
streamlit
pandas
google-cloud-logging
google-cloud-bigquery
google.cloud
google-cloud-secret-manager==2.0.0
google.cloud.logging
rich
numpy
millify
plost
plotly
streamlit_option_menu
pyinstrument
db-dtypes
st-pages
